# Add endpoint builder to define EDS dependencies