Goals:
* Scope down the set of inputs to the EDS pipeline so it doesn't depend
on proxy
* Compute some things up front to reduce recomputation
* Define a key that can be used for caching EDS responses (future PR)

```
$ benchstat /tmp/{old,new}
name                         old time/op    new time/op    delta
EndpointGeneration/1/100-6      139µs ± 2%     119µs ± 2%  -14.38%  (p=0.008 n=5+5)
EndpointGeneration/10/10-6     67.4µs ± 2%    64.9µs ± 2%   -3.67%  (p=0.016 n=5+5)
EndpointGeneration/100/10-6     472µs ± 4%     472µs ± 6%     ~     (p=1.000 n=5+5)
EndpointGeneration/1000/1-6     708µs ± 6%     684µs ± 9%     ~     (p=0.310 n=5+5)

name                         old alloc/op   new alloc/op   delta
EndpointGeneration/1/100-6     63.7kB ± 0%    57.3kB ± 0%  -10.05%  (p=0.008 n=5+5)
EndpointGeneration/10/10-6     10.7kB ± 0%    10.1kB ± 0%   -5.96%  (p=0.008 n=5+5)
EndpointGeneration/100/10-6    30.5kB ± 0%    29.8kB ± 1%   -2.32%  (p=0.008 n=5+5)
EndpointGeneration/1000/1-6    36.1kB ± 0%    34.9kB ± 6%   -3.38%  (p=0.032 n=5+5)

name                         old allocs/op  new allocs/op  delta
EndpointGeneration/1/100-6      1.06k ± 0%     0.96k ± 0%   -9.43%  (p=0.008 n=5+5)
EndpointGeneration/10/10-6        155 ± 0%       145 ± 0%   -6.45%  (p=0.008 n=5+5)
EndpointGeneration/100/10-6       160 ± 0%       150 ± 2%   -6.50%  (p=0.016 n=4+5)
EndpointGeneration/1000/1-6      22.0 ± 0%      22.4 ±16%     ~     (p=0.643 n=5+5)
```