@istio-testing 2020-07-17
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
release-notes_istio | d9969165ce5730e2cb1db213332df300b3314722 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25598/release-notes_istio/39) | `/test release-notes_istio`
lint_istio | d9969165ce5730e2cb1db213332df300b3314722 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25598/lint_istio/16221) | `/test lint_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->