@istio-testing 2020-07-20
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
benchmark_istio | a18928a7ac48deb10afa8e3d5d082e497ff4626c | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/25598/benchmark_istio/9) | `/test benchmark_istio`
integ-multicluster-k8s-tests_istio | a18928a7ac48deb10afa8e3d5d082e497ff4626c | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/25598/integ-multicluster-k8s-tests_istio/7080) | `/test integ-multicluster-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->